+++ iocore/eventsystem/I_IOBuffer.h
@@ -346,7 +346,7 @@ public:
   int64_t
   size()
   {
-    return (int64_t)(_end - _start);
+    return _end - _start;
   }
 
   /**
@@ -359,7 +359,7 @@ public:
   int64_t
   read_avail()
   {
-    return (int64_t)(_end - _start);
+    return _end - _start;
   }
 
   /**
@@ -371,7 +371,7 @@ public:
   int64_t
   write_avail()
   {
-    return (int64_t)(_buf_end - _end);
+    return _buf_end - _end;
   }
 
   /**
